+++ rust/numaflow-core/src/pipeline.rs
@@ -23,9 +23,9 @@ use crate::pipeline::isb::jetstream::reader::JetStreamReader;
 use crate::pipeline::isb::jetstream::writer::JetstreamWriter;
 use crate::pipeline::pipeline::isb::BufferReaderConfig;
 use crate::reduce::pbq::PBQBuilder;
-use crate::reduce::pnf::aligned::reducer::AlignedReducer;
-use crate::reduce::pnf::aligned::windower::fixed::FixedWindowManager;
-use crate::reduce::pnf::aligned::windower::sliding::SlidingWindowManager;
+use crate::reduce::reducer::aligned::reducer::AlignedReducer;
+use crate::reduce::reducer::aligned::windower::fixed::FixedWindowManager;
+use crate::reduce::reducer::aligned::windower::sliding::SlidingWindowManager;
 use crate::reduce::wal::segment::WalType;
 use crate::reduce::wal::segment::append::AppendOnlyWal;
 use crate::reduce::wal::segment::compactor::{Compactor, WindowKind};
