+++ rust/numaflow-core/src/reduce.rs
@@ -11,4 +11,4 @@ mod error;
 pub(crate) mod pbq;
 
 /// Process and Forward (PNF) for processing the reduce data and forwarding it to the next vertex.
-pub(crate) mod pnf;
+pub(crate) mod reducer;
