+++ rust/numaflow-core/src/reduce/reducer/aligned/windower.rs
